pyproject.toml
[tool.poetry]
name = "testpilot"
version = "0.1.0"
description = ""
authors = ["jllewis11 <jllewis11@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
testpilot = "testpilot.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.1"
together = "^0.2.11"
openai = "^1.13.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
